feat(difs): Propagate download info to dif candidates
This commit re-arranges the objects actor among multiple modules to
make it more manageable.  No functional change.
